+++ libs/react-components/community/src/components/card/card.stories.tsx
@@ -14,7 +14,7 @@ import CardHeader, { CardHeaderProps } from './card-header/card-header';
 import CardNotification, { CardNotificationProps } from './card-notification/card-notification';
 
 export default {
-  title: 'Community-components/Card',
+  title: 'Community/Card',
   component: Card,
   subcomponents: { CardContent, CardHeader, CardNotification },
 } as Meta;