+++ libs/react-components/community/src/components/form/check/check.stories.tsx
@@ -11,7 +11,7 @@ import Check, { CheckProps } from './check';
  */
 const meta: Meta<typeof Check> = {
   component: Check,
-  title: 'Community-components/Form/Check',
+  title: 'Community/Form/Check',
 };
 
 export default meta;