+++ libs/react-components/community/src/components/form/hidden-field/hidden-field.stories.tsx
@@ -7,7 +7,7 @@ import HiddenField from './hidden-field';
 
 const meta: Meta<typeof HiddenField> = {
   component: HiddenField,
-  title: 'Community-components/Form/HiddenField',
+  title: 'Community/Form/HiddenField',
   argTypes: {
     content: { control: false },
     fieldType: { control: false },