+++ libs/react-components/community/src/components/form/pickers/datepicker/datepicker.stories.tsx
@@ -8,7 +8,7 @@ import DatePicker, { DatePickerProps } from './datepicker';
 
 const meta: Meta<typeof DatePicker> = {
   component: DatePicker,
-  title: 'Community-components/Form/Pickers/DatePicker',
+  title: 'Community/Form/Pickers/DatePicker',
 };
 
 export default meta;