+++ libs/react-components/community/src/components/layout/header/components/header-role/header-role.stories.tsx
@@ -12,7 +12,7 @@ import RoleSelection from './header-role';
  */
 const meta: Meta<typeof RoleSelection> = {
   component: RoleSelection,
-  title: 'Community-components/Layout/Header/HeaderRole',
+  title: 'Community/Layout/Header/HeaderRole',
 };
 
 export default meta;