+++ libs/react-components/community/src/components/placeholder/placeholder.stories.tsx
@@ -8,7 +8,7 @@ import Placeholder from './placeholder';
 
 const meta: Meta<typeof Placeholder> = {
   component: Placeholder,
-  title: 'Community-components/Placeholder',
+  title: 'Community/Placeholder',
 };
 
 export default meta;