+++ libs/react-components/community/src/components/typography/heading/heading.stories.tsx
@@ -9,7 +9,7 @@ import Heading, { HeadingProps } from './heading';
  */
 const meta: Meta<typeof Heading> = {
   component: Heading,
-  title: 'Community-components/Typography/Heading',
+  title: 'Community/Typography/Heading',
 };
 
 export default meta;