+++ libs/react-components/tedi/src/index.ts
@@ -1,6 +1,7 @@
 export * from './components/spinner/spinner';
 export * from './components/typography/text/text';
 export * from './components/typography/heading/heading';
+export * from './components/icon/icon';
 
 export * from './helpers';